third_party/blink/web_tests/external/wpt/infrastructure/testdriver/actions/penPointerEventProperties-expected.txt

This is a testharness.js-based test.
FAIL TestDriver actions: pointerevent properties of pen type assert_equals: expected 0.36 but got 0.5
Harness: the test ran to completion.
